Move yet more stuff out of lib.php
We don't want any globally available un-namespaced funcs. Partly because that's
just a woolly practice, and partly because psalm can't understand that they
actually exist resulting it loads of entries in our psalm baseline.
