+++ gatherling/Views/Pages/EventForm.php
@@ -17,6 +17,7 @@ use Gatherling\Views\Components\SeasonDropMenu;
 use Gatherling\Views\Components\SeriesDropMenu;
 use Gatherling\Views\Components\TimeDropMenu;
 
+use function Gatherling\Helpers\getObjectVarsCamelCase;
 use function Safe\preg_match;
 
 class EventForm extends EventFrame
