+++ phpstan-baseline.neon
@@ -21,6 +21,16 @@ parameters:
 			path: gatherling/Helpers/cards.php
 
 		-
+			message: "#^Function Gatherling\\\\Helpers\\\\arrayMapRecursive\\(\\) should return array\\<string, mixed\\> but returns array\\<int\\|string, mixed\\>\\.$#"
+			count: 1
+			path: gatherling/Helpers/templates.php
+
+		-
+			message: "#^Parameter \\#2 \\$arr of function Gatherling\\\\Helpers\\\\arrayMapRecursive expects array\\<string, mixed\\>, array given\\.$#"
+			count: 1
+			path: gatherling/Helpers/templates.php
+
+		-
 			message: "#^Cannot access property \\$data on mixed\\.$#"
 			count: 1
 			path: gatherling/Models/CardSet.php
@@ -947,7 +957,7 @@ parameters:
 
 		-
 			message: "#^Cannot cast mixed to int\\.$#"
-			count: 2
+			count: 1
 			path: gatherling/admincp.php
 
 		-
@@ -1119,16 +1129,6 @@ parameters:
 			message: "#^Part \\$_POST\\['action'\\] \\(mixed\\) of encapsed string cannot be cast to string\\.$#"
 			count: 1
 			path: gatherling/formatcp.php
-
-		-
-			message: "#^Function arrayMapRecursive\\(\\) should return array\\<string, mixed\\> but returns array\\<int\\|string, mixed\\>\\.$#"
-			count: 1
-			path: gatherling/lib.php
-
-		-
-			message: "#^Parameter \\#2 \\$arr of function arrayMapRecursive expects array\\<string, mixed\\>, array given\\.$#"
-			count: 1
-			path: gatherling/lib.php
 
 		-
 			message: "#^Property Gatherling\\\\Models\\\\Player\\:\\:\\$emailAddress \\(string\\|null\\) does not accept mixed\\.$#"
